Remove separate colors file, fix disposal
